wip: consolidate based on concerns
