+++ src/index.ts
@@ -14,8 +14,8 @@ export * from '@/Kernel';
 // Routing
 export * from '@/routing/deprecated-decorator/route';
 export type { HttpMethod } from '@/routing/verb/http-method.type';
-export type { RouteOptions } from '@/routing/route-options';
-export type { RouterMethod } from '@/routing/router-method.type';
+export type { RouteOptions } from '@/routing/route/route-options';
+export type { RouterMethod } from '@/routing/route/router-method.type';
 /**
  * @deprecated Use `Route` from `@koala-ts/framework/routing` instead.
  */
