+++ src/routing/index.ts
@@ -1,5 +1,5 @@
-export type { RouteSource } from '@/routing/source/route-source';
-export { Route } from './route';
-export { createPathFor } from './helpers/create-path-for';
-export { RouteGroup } from './helpers/route-group';
-export { Any, Delete, Get, Head, Options, Patch, Post, Put } from './helpers/http-verb-helpers';
+export type { RouteSource } from '@/routing/route/source/route-source';
+export { Route } from './route/route';
+export { createPathFor } from './path/create-path-for';
+export { RouteGroup } from './group/route-group';
+export { Any, Delete, Get, Head, Options, Patch, Post, Put } from './verb/http-verb-helpers';
