+++ src/routing/route/registration/expand-route-definitions.ts
@@ -1,7 +1,7 @@
-import { koaBody } from 'koa-body';
+import type { RouteDefinition } from '@/routing/route/definition/route-definition';
+import type { RouteRegistration } from '@/routing/route/registration/route-registration';
 import type { Request } from 'koa';
-import type { RouteDefinition } from '@/routing/definition/route-definition';
-import type { RouteRegistration } from '@/routing/registration/route-registration';
+import { koaBody } from 'koa-body';
 
 export function expandRouteDefinitions<TRequest extends Request>(
   routes: RouteDefinition<TRequest>[],
