+++ src/routing/route/source/resolve-path-template.ts
